move envelope fetch to Month.js
